\name{drug_response_w12}
\alias{drug_response_w12}
\docType{data}
\title{drug_response_w12 data}
\description{
A matrix containing drug response values (IC50 values) obtained from patients in Waves 1 and 2 of the BeatAML2 cohort. This dataset serves as a toy example for demonstrating the functionality of the \code{trainTree}, \code{predictTree}, and \code{niceTree} functions.
}

\usage{data("drug_response_w12")}
\format{
  The format is:
  num [1:247, 1:119] 2.710983 2.8755433 3.4390103 2.6527257...
}
\examples{
data(drug_response_w12)
}
\keyword{datasets}
