\name{drug_response_w34}
\alias{drug_response_w34}
\docType{data}
\title{drug_response_w34 data}
\description{A matrix containing drug response values (IC50 values) from patients retrieved from waves 3 and 4. Used as a toy example in \code{trainTree} and \code{predictTree} and \code{niceTree}
}
\usage{data("drug_response_w34")}
\format{
  The format is:
  num [1:247, 1:119] 3.4156359 3.2345985 3.1836058 3.7874252...
}
\examples{
data(drug_response_w34)
}
\keyword{datasets}