\name{expression_w12}
\alias{expression_w12}
\docType{data}
\title{expression_w12 Data Set}
\description{A dataframe containing gene expression values (with different types of genes) from patients retrieved from waves 1 and 2. Used as a toy example in \code{trainTree} and \code{predictTree} and \code{niceTree}
}
\usage{data("expression_w12")}
\format{
  A data frame with 22843 observations on the following 247 variables, 
  where there are 247 patients from the waves and 22843 different genes   are considered.
  
  The format is:
  22843 obs. 247 variables 
}
\examples{
data(expression_w12)
}
\keyword{datasets}
