\name{expression_w34}
\alias{expression_w34}
\docType{data}
\title{expression_w34 Data Set}
\description{A dataframe containing gene expression values obtained from patients in Waves 3 and 4 of the BeatAML cohort. This dataset serves as a toy example for demonstrating the functionality of the \code{trainTree}, \code{predictTree}, and \code{niceTree} functions.
}
\usage{data("expression_w34")}
\format{A dataframe consisting of 142 rows and 1000 columns, where each row represents a different patient and each column corresponds to the expression levels of a specific gene. The entries in the data frame are floating-point values, indicating the gene expression levels measured for each patient.
}
\examples{
data(expression_w34)
}
\keyword{datasets}
