% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niceTree.R
\name{niceTree}
\alias{niceTree}
\title{niceTree function}
\usage{
niceTree(
  tree,
  folder = NULL,
  colors = c("", "#367592", "#39A7AE", "#96D6B6", "#FDE5B0", "#F3908B", "#E36192",
    "#8E4884", "#A83333"),
  fontname = "Roboto",
  fontstyle = "plain",
  shape = "diamond",
  output_format = "png"
)
}
\arguments{
\item{tree}{A party of the trained tree with the treatments assigned to each node.}

\item{folder}{Directory to save the image}

\item{colors}{Selection of colors for the boxes}

\item{fontname}{The font name}

\item{fontstyle}{The font style}

\item{shape}{format of the boxes of the different genes}

\item{output_format}{image format}
}
\value{
(Invisibly) returns a list. The representation of the tree in the command window and the plot of the tree
}
\description{
A graphical display of the tree. It can also be saved as an image in the selected directory.
}
\details{
\itemize{
 \item The user has already a defined style for the plot, the parameters
       are previously set if the parameters are not modified when calling
       niceTree
}
}
\examples{

\dontrun{
  #This first example is a basic example of how to perform niceTree: 
  
  data(DataODT.rda)
  ODTmut <- trainTree(PatientData = mutations_w12,
  PatientSensitivity=drug_response_w12, minbucket =10)
  niceTree(ODTMut,folder="")

  
  #The next example, is the same as the first
  # one but, plotting the tree trained for gene expressions. 
  
  data(DataODT.rda)
  ODTExp <- trainTree(PatientData=expression_w34,
  PatientSensitivity=drug_response_w34, minbucket = 20)
  niceTree(ODTExp,folder="")
  
  
 }
 
}
