\name{drug_response_w34}
\alias{drug_response_w34}
\docType{data}
\title{drug_response_w34 data}
\description{
A matrix containing drug response values (IC50 values) obtained from patients in Waves 3 and 4 of the BeatAML2 cohort. This dataset serves as a toy example for demonstrating the functionality of the \code{trainTree}, \code{predictTree}, and \code{niceTree} functions.
}

\usage{data("drug_response_w34")}
\format{
  The format is:
  num [1:142, 1:119] 3.4156359 3.2345985 3.1836058 3.7874252...
}
\examples{
data(drug_response_w34)
}
\keyword{datasets}