\name{expression_w34}
\alias{expression_w34}
\docType{data}
\title{expression_w34 Data Set}
\description{A dataframe containing gene expression values (with different types of genes) from patients retrieved from waves 3 and 4. Used as a toy example in \code{trainTree} and \code{predictTree} and \code{niceTree}
}
\usage{data("expression_w34")}
\format{
  A data frame with 22843 observations on the following 142 variables, 
  where there are 142 patients from the waves and 22843 different genes   are considered.
  
  The format is:
  22843 obs. 142 variables 
}
\examples{
data(expression_w34)
}
\keyword{datasets}
