% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxFunctions.R
\name{InternalFunctions}
\alias{InternalFunctions}
\alias{findsplitExp}
\title{ODT Internal Functions}
\usage{
findsplitExp(
  PatientSensitivity,
  PatientData,
  minimum = 1,
  weights = NULL,
  verbose = F
)
}
\value{
Internal outputs
}
\description{
Internal functions used by trainTree in the different
steps of the algorithm
}
\keyword{internal}
