% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainTree.R
\name{trainTree}
\alias{trainTree}
\title{trainTree Function}
\usage{
trainTree(PatientData, PatientSensitivity, minbucket = 20)
}
\arguments{
\item{PatientData}{A matrix representing either gene expression data or mutation information.
This matrix can be binary (for mutational data) or continuous (for expression data).}

\item{PatientSensitivity}{A numeric vector indicating drug response values.
Higher values suggest greater drug resistance and, consequently, less sensitivity.
This vector can represent various measures of drug response, such as IC50,
area under the drug response curve (AUC), or other relevant metrics.
Depending on the interpretation of the sensitivity response, the user may need to adjust the sign of the data.}

\item{minbucket}{An integer specifying the minimum number of patients required in a node to allow for a split.}
}
\value{
An object of class 'party' representing the trained decision tree, with the assigned treatments for each node.
}
\description{
This function trains a decision tree model based on patient data, which can either be gene expression levels or a binary matrix indicating mutations.
}
\examples{

\dontrun{
  # Basic example of using the trainTree function with mutational data
  data(DataODT.rda)
  ODTmut <- trainTree(PatientData = mutations_w12, 
                      PatientSensitivity = drug_response_w12,
                      minbucket = 10)
  plot(ODTmut)

  # Example using gene expression data instead
  data(DataODT.rda)
  ODTExp <- trainTree(PatientData = expression_w34,
                      PatientSensitivity = drug_response_w34,
                      minbucket = 20)
  plot(ODTExp)
}

}
